refactor: extract sealed operations from server.go
